criação dos end-points alteraEstadoBicicleta e IntegraBicicletaNaRede

+++ SistemaBicicletario/src/main/java/com/mariaClara/SistemaBicicletario/repository/BicicletaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BicicletaRepository extends JpaRepository<BicicletaEntity, Long>{
+public interface BicicletaRepository extends JpaRepository<BicicletaEntity, Integer>{
 }
